+++ iroh-content/Cargo.toml
@@ -2,13 +2,15 @@
 name = "iroh-content"
 version = "0.1.3"
 edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+authors = ["dignifiedquire <me@dignifiedquire.com>"]
+license = "Apache-2.0/MIT"
+repository = "https://github.com/n0-computer/iroh"
+description = "Implementation of unixfs for iroh"
+rust-version = "1.63"
 
 [dependencies]
 anyhow = "1"
 async-trait = "0.1.53"
-async-stream = "0.3.3"
 bytes = "1.1.0"
 cid = "0.9"
 futures = "0.3.21"
@@ -23,7 +25,6 @@ libipld = "0.15.0"
 iroh-util = { version = "0.1.3", path = "../iroh-util", default-features = false }
 tokio = { version = "1", features = ["fs"] }
 async-channel = "1.7.1"
-iroh-metrics = { version = "0.1.3", path = "../iroh-metrics", default-features = false, features = ["resolver", "gateway"] }
 base64 = "0.13.1"
 unsigned-varint = "0.7.1"
 num_enum = "0.5.7"
