+++ iroh-one/src/uds.rs
@@ -3,10 +3,7 @@
 use axum::extract::connect_info;
 use axum::{Router, Server};
 use futures::ready;
-use hyper::{
-    client::connect::{Connected, Connection},
-    server::accept::Accept,
-};
+use hyper::server::accept::Accept;
 use iroh_content::content_loader::ContentLoader;
 use iroh_gateway::{core::State, handlers::get_app_routes};
 use std::path::PathBuf;
